good version
Note, this version meets my initial goal of playing well in little memory.  Without a transposition table, it still runs well enough to beat TSCP and match or possibly beat Ed Schröder's alteration of TSCP (my run against it was short so I'm not sure).   So with a little tweaking this will run on the arduino